rename: auth 관련 기능 Auth~로 이동 및 수정
