--- src/main/java/com/fithub/fithubbackend/domain/user/repository/SignUpRepository.java
@@ -1,11 +1,0 @@
-package com.fithub.fithubbackend.domain.user.repository;
-
-import com.fithub.fithubbackend.domain.user.domain.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import java.util.Optional;
-
-public interface SignUpRepository extends JpaRepository<User,Long> {
-    Optional<User> findByNickname(String nickname);
-    Optional<User> findByEmail(String email);
-}
