+++ src/main/java/com/fithub/fithubbackend/domain/user/repository/UserRepository.java
@@ -6,6 +6,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
 
 public interface UserRepository extends JpaRepository<User, Long> {
-
+    Optional<User> findByNickname(String nickname);
     Optional<User> findByEmail(String email);
 }
